SettingsMenu cleanup & alert ref fix
