+++ src/components/[auction]/SettingsMenu/index.ts
@@ -1,0 +1,3 @@
+import SettingsMenu from "./SettingsMenu"
+
+export default SettingsMenu
